Push down responsibility set annotation
