+++ app/src/main/admin/components/Touchstones/SingleTouchstoneVersion/ResponsibilitySet.tsx
@@ -6,21 +6,40 @@ import {ResponsibilityList} from "./ResponsibilityList";
 import {AnnotatedResponsibilitySet} from "../../../models/AnnotatedResponsibility";
 import {adminTouchstoneActionCreators} from "../../../actions/adminTouchstoneActionCreators";
 import {compose} from "recompose";
+import {ResponsibilitySetWithComments, ResponsibilitySetWithExpectations} from "../../../../shared/models/Generated";
 
 interface ResponsibilitySetProps {
-    responsibilitySet: AnnotatedResponsibilitySet;
+    responsibilitySet: ResponsibilitySetWithExpectations;
     canReviewResponsibilities: boolean;
+    responsibilityComments: ResponsibilitySetWithComments[];
     setCurrentTouchstoneResponsibilitySet: (responsibilitySet: AnnotatedResponsibilitySet) => void;
 }
 
-class ResponsibilitySetComponent extends React.Component<ResponsibilitySetProps> {
+interface ResponsibilitySetState {
+    annotatedResponsibilitySet: AnnotatedResponsibilitySet;
+}
+
+class ResponsibilitySetComponent extends React.Component<ResponsibilitySetProps, ResponsibilitySetState> {
+
+    constructor(props: ResponsibilitySetProps) {
+        super(props);
+        this.state = {
+            annotatedResponsibilitySet: toAnnotatedResponsibilitySet(props.responsibilitySet, props.responsibilityComments)
+        };
+    }
 
     handleClick() {
-        this.props.setCurrentTouchstoneResponsibilitySet(this.props.responsibilitySet);
+        this.props.setCurrentTouchstoneResponsibilitySet(this.state.annotatedResponsibilitySet);
+    }
+
+    componentWillReceiveProps(nextProps: Readonly<ResponsibilitySetProps>) {
+        this.setState({
+            annotatedResponsibilitySet: toAnnotatedResponsibilitySet(nextProps.responsibilitySet, nextProps.responsibilityComments)
+        });
     }
 
     render(): JSX.Element {
-        const responsibilitySet = this.props.responsibilitySet;
+        const responsibilitySet = this.state.annotatedResponsibilitySet;
         return <div>
             <h4>{responsibilitySet.modelling_group_id} (<span>{responsibilitySet.status}</span>)</h4>
             {this.props.canReviewResponsibilities &&
@@ -41,9 +60,30 @@ class ResponsibilitySetComponent extends React.Component<ResponsibilitySetProps>
 
 }
 
+function toAnnotatedResponsibilitySet(
+    responsibilitySet: ResponsibilitySetWithExpectations,
+    responsibilityComments: ResponsibilitySetWithComments[]
+): AnnotatedResponsibilitySet {
+    const responsibilitySetWithComments = responsibilityComments
+        .find(e => e.modelling_group_id === responsibilitySet.modelling_group_id)
+    return {
+        comment: responsibilitySetWithComments && responsibilitySetWithComments.comment,
+        ...responsibilitySet,
+        responsibilities: responsibilitySet.responsibilities.map(r => (
+            {
+                modellingGroup: responsibilitySet.modelling_group_id,
+                comment: responsibilitySetWithComments && responsibilitySetWithComments.responsibilities
+                    .find(e => e.scenario_id === r.scenario.id).comment,
+                ...r
+            }
+        ))
+    };
+}
+
 const mapStateToProps = (state: AdminAppState): Partial<ResponsibilitySetProps> => {
     return {
-        canReviewResponsibilities: state.auth.canReviewResponsibilities
+        canReviewResponsibilities: state.auth.canReviewResponsibilities,
+        responsibilityComments: state.touchstones.currentResponsibilityComments
     }
 };
